# CUDA: remove DMMV, consolidate F16 mult mat vec