@ggerganov 2024-11-15
On RTX 3090, without FA is faster than with FA for tg8192. Is this expected?