@JohannesGaessler 2024-11-15
I forgot to mention: this is because the 1b model has a head size of 64 (vs. 128 in the larger LLaMA variants) and the CUDA FA kernel is comparatively poorly optimized for that head size.