@JohannesGaessler 2024-11-15
That's one way to stop the singularity from happening lol. What's funny is that it counts the deletion of `dmmv.cu` as a successful review even though it's impossible for it to read the code.