@slaren 2024-11-16
Incidentally this also seems to improve performance with the `stories260k.gguf` F32 model, but I am not sure why. The model is F32, so it shouldn't use this path, but maybe it is used for the attention. I was also wondering if it may be worth also making a version of this kernel for F32, since currently it only supports F16.

| Model                      | Test   |   t/s master |   t/s cuda-mmv-5 |   Speedup |
|:---------------------------|:-------|-------------:|-----------------:|----------:|
| llama ?B all F32 (guessed) | pp512  |    278910.28 |        278139.21 |      1.00 |
| llama ?B all F32 (guessed) | tg1024 |      2522.71 |          2882.23 |      1.14 |