@JohannesGaessler 2024-11-17
For a FP32 model the only difference is in the attention, particularly the calculation of KQV. My expectation for FP32 would be that cuBLAS GEMM would be faster (but I had also expected cuBLAS GEMM to be faster for FP16).