@LostRuins 2025-02-09
Hi @JohannesGaessler and @slaren , sorry for the necropost but this appears to be the PR that triggers a regression issue for old GPUs. 

Specifically, it seems like now we are required to have `FP16_AVAILABLE` in order to use `mul_mat_vec`

https://github.com/ggerganov/llama.cpp/blob/19d3c8293b1f61acbe2dab1d49a17950fd788a4a/ggml/src/ggml-cuda/mmv.cu#L47-L58

However, there doesn't actually seem to be anything preventing someone with an old GPU with **no FP16 support** from calling `mul_mat_vec`, for example maybe when `src0->ne[1]` is 400.

https://github.com/ggerganov/llama.cpp/blob/19d3c8293b1f61acbe2dab1d49a17950fd788a4a/ggml/src/ggml-cuda/ggml-cuda.cu#L1847-L1849

https://github.com/ggerganov/llama.cpp/blob/19d3c8293b1f61acbe2dab1d49a17950fd788a4a/ggml/src/ggml-cuda/ggml-cuda.cu#L1888-L1891

This results in a `ERROR: CUDA kernel mul_mat_vec has no device code compatible with CUDA arch 520. ` on old GPUs.

Perhaps we need to guard against old GPUs entering that branch?