@JohannesGaessler 2025-02-09
Thank you for bringing this to my attention, should be fixed by https://github.com/ggerganov/llama.cpp/pull/11775 .